rename: selfTest -> reviewSelf로 상태명 변경 및 서버와 state 일치 작업
